Move process event consumer imported only by system-probe
